requirements.txt
telethon
pydantic-settings
python-dotenv
cryptg
aiohttp
SpeechRecognition
pydub
Pillow
